+++ day16_02/app01/views/number.py
@@ -1,3 +1,4 @@
+from django.http import JsonResponse
 from django.shortcuts import render, redirect
 from app01 import models
 from django.views.decorators.csrf import csrf_exempt
@@ -55,5 +56,9 @@ def number_edit(request, nid):
 
 
 def number_delete(request, nid):
+    exist = models.PhoneNumber.objects.filter(id=nid).exists()
+    if not exist:
+        return JsonResponse({"status": False, "error":"此数据不存在"})
+
     models.PhoneNumber.objects.filter(id=nid).delete()
-    return redirect('/number/list/')
+    return JsonResponse({"status": True})
